Organize source directory, classes and their mocks live together
