+++ app/src/main/java/waiter/ClientConnection/Connection.java
@@ -1,4 +1,6 @@
-package waiter;
+package waiter.ClientConnection;
+
+import waiter.Message;
 
 import java.io.IOException;
