+++ app/src/main/java/waiter/EchoProtocol/EchoProtocol.java
@@ -1,4 +1,6 @@
-package waiter;
+package waiter.EchoProtocol;
+
+import waiter.Message;
 
 public class EchoProtocol implements Protocol {
     public Message serve(Message fromClient) {
