+++ app/src/main/java/waiter/EchoProtocol/Protocol.java
@@ -1,4 +1,6 @@
-package waiter;
+package waiter.EchoProtocol;
+
+import waiter.Message;
 
 public interface Protocol {
     Message serve(Message fromClient);
