+++ app/src/main/java/waiter/EchoProtocol/mock/EchoProtocolMock.java
@@ -1,4 +1,7 @@
-package waiter;
+package waiter.EchoProtocol.mock;
+
+import waiter.Message;
+import waiter.EchoProtocol.Protocol;
 
 public class EchoProtocolMock implements Protocol {
     public Message serve(Message mockMessage) {
