+++ app/src/main/java/waiter/Listener/Listener.java
@@ -1,4 +1,7 @@
-package waiter;
+package waiter.Listener;
+
+import waiter.ClientConnection.ClientConnection;
+import waiter.ClientConnection.Connection;
 
 import java.io.IOException;
 import java.net.ServerSocket;
@@ -6,7 +9,7 @@ import java.net.ServerSocket;
 public class Listener implements Awaiter {
     ServerSocket serverSocket;
 
-    Listener(ServerSocket serverSocket) {
+    public Listener(ServerSocket serverSocket) {
         this.serverSocket = serverSocket;
     }
