+++ app/src/main/java/waiter/Listener/mock/ListenerMock.java
@@ -1,10 +1,14 @@
-package waiter;
+package waiter.Listener.mock;
+
+import waiter.Listener.Awaiter;
+import waiter.ClientConnection.Connection;
+import waiter.ClientConnection.mock.ClientConnectionMock;
 
 public class ListenerMock implements Awaiter {
 
     String userInput;
 
-    ListenerMock(String userInput) {
+    public ListenerMock(String userInput) {
         this.userInput = userInput;
     }
