+++ app/src/main/java/waiter/Messenger/mock/MessengerMock.java
@@ -1,8 +1,11 @@
-package waiter;
+package waiter.Messenger.mock;
+
+import waiter.ClientConnection.Connection;
+import waiter.Messenger.Transporter;
 
 public class MessengerMock implements Transporter {
 
-    Connection calledWith;
+    public Connection calledWith;
 
     public void transport(Connection clientConnectionMock) {
         this.calledWith = clientConnectionMock;
